+++ xcube_smos/catalog/__init__.py
@@ -1,4 +1,4 @@
 from .base import AbstractSmosCatalog
 from .base import DatasetOpener
-from .index import SmosIndexCatalog
 from .direct import SmosDirectCatalog
+from .stac import SmosStacCatalog