+++ xcube_smos/catalog/base.py
@@ -1,5 +1,5 @@
 # The MIT License (MIT)
-# Copyright (c) 2023 by the xcube development team and contributors
+# Copyright (c) 2023-2024 by the xcube development team and contributors
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -22,11 +22,13 @@
 import abc
 from typing import Dict, Any, Optional, Tuple, List
 
-from xcube_smos.catalog.producttype import ProductTypeLike
+import pandas as pd
+
 from ..utils import NotSerializable
+from .producttype import ProductTypeLike
 from .types import DatasetOpener
 from .types import DatasetRecord
-from .types import AcceptRecord
+from .types import DatasetFilter
 
 
 class AbstractSmosCatalog(NotSerializable, abc.ABC):
@@ -94,13 +96,15 @@ class AbstractSmosCatalog(NotSerializable, abc.ABC):
 
     # noinspection PyMethodMayBeStatic
     def resolve_path(self, dataset_path: str) -> str:
-        """Resolve the given path returned by ``find_datasets()``.
+        """Resolve the given path returned by `find_datasets()`.
 
-        :param dataset_path: Unresolved dataset path as returned by
-            ``find_datasets()``.
-        :return: A resolved dataset path that is passed to the
-            (static) ``open_dataset`` function returned by the
-            ``get_dataset_opener()`` method.
+        Args:
+            dataset_path: Unresolved dataset path as returned by
+                `find_datasets()`.
+        Returns:
+            A resolved dataset path that is passed to the
+            (static) `open_dataset` function returned by the
+            `get_dataset_opener()` method.
         """
         return dataset_path
 
@@ -108,18 +112,23 @@ class AbstractSmosCatalog(NotSerializable, abc.ABC):
     def find_datasets(
         self,
         product_type: ProductTypeLike,
-        time_range: Tuple[Optional[str], Optional[str]],
-        accept_record: Optional[AcceptRecord] = None,
+        time_range: Tuple[pd.Timestamp, pd.Timestamp],
+        dataset_filter: Optional[DatasetFilter] = None,
+        **query_parameters
     ) -> List[DatasetRecord]:
         """Find SMOS L2 datasets in the given *time_range*.
 
-        :param product_type: SMOS product type
-        :param time_range: Time range (from, to) ISO format, UTC
-        :param accept_record: An optional dataset filter function,
-            that receives a dataset record.
-        :return: List of dataset records of the form
+        Args:
+            product_type: SMOS product type
+            time_range: Time range (from, to) ISO format, UTC
+            dataset_filter: An optional filter function,
+                which receives catalog-specific product information.
+            query_parameters: Optional catalog-specific
+                query parameters.
+        Returns:
+            List of dataset records of the form
             (*dataset_path*, *start*, *stop*), where *dataset_path*
             is yet unresolved and *start*, *stop* represent the observation
-            time range using "compact" datetime format,
-            e.g., ``"20230503103546"``.
+            time range using ISO datetime format,
+            e.g., `"2023-05-03T10:35:46.000Z"`.
         """