+++ xcube_smos/catalog/direct.py
@@ -1,5 +1,5 @@
 # The MIT License (MIT)
-# Copyright (c) 2023 by the xcube development team and contributors
+# Copyright (c) 2023-2024 by the xcube development team and contributors
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -20,17 +20,21 @@
 # DEALINGS IN THE SOFTWARE.
 
 import atexit
+import warnings
 from functools import cached_property
 import logging
 import os
 from pathlib import Path
+import re
 import shutil
 import tempfile
-from typing import Union, Dict, Any, Optional, Tuple, List, Set, Iterable
+from typing import Dict, Any, Set, Iterable, Union, Tuple, Optional, List, Callable
 
 import fsspec
+import pandas as pd
 import xarray as xr
 
+from ..constants import COMPACT_DATETIME_FORMAT
 from ..constants import DEFAULT_ARCHIVE_URL
 from ..constants import DEFAULT_STORAGE_OPTIONS
 from ..constants import OS_VAR_NAMES
@@ -40,8 +44,12 @@ from .producttype import ProductType
 from .producttype import ProductTypeLike
 from .types import DatasetOpener
 from .types import DatasetRecord
-from .types import AcceptRecord
+from .types import DatasetFilter
 
+
+GetFilesForPath = Callable[[str], Iterable[str]]
+
+_ONE_DAY = pd.Timedelta(1, unit="days")
 
 LOG = logging.getLogger("xcube-smos")
 
@@ -93,12 +101,21 @@ class SmosDirectCatalog(AbstractSmosCatalog):
     def find_datasets(
         self,
         product_type: ProductTypeLike,
-        time_range: Tuple[Optional[str], Optional[str]],
-        accept_record: Optional[AcceptRecord] = None,
+        time_range: Tuple[pd.Timestamp, pd.Timestamp],
+        dataset_filter: Optional[DatasetFilter] = None,
+        **query_parameters,
     ) -> List[DatasetRecord]:
+        if query_parameters:
+            warnings.warn(
+                f"Additional parameter(s) not understood:"
+                f" {', '.join(query_parameters.keys())}"
+            )
         product_type = ProductType.normalize(product_type)
-        return product_type.find_files_for_time_range(
-            time_range, self._get_files_for_path, accept_record=accept_record
+        return find_files_for_time_range(
+            product_type,
+            time_range,
+            self._get_files_for_path,
+            dataset_filter=dataset_filter,
         )
 
     def _get_files_for_path(self, path: str) -> Iterable[str]:
@@ -155,7 +172,7 @@ def open_dataset(
     else:
         var_names = None
 
-    remote_fs = fsspec.filesystem(source_protocol, **source_storage_options)
+    remote_fs = fsspec.filesystem(source_protocol, **(source_storage_options or {}))
     if not cache_path:
         if open_dataset_kwargs.get("engine") == "h5netcdf":
             LOG.debug("Opening dataset directly from %s", source_file)
@@ -198,3 +215,102 @@ def filter_dataset(ds: xr.Dataset, var_names: Set[str]) -> xr.Dataset:
         k[len(key_prefix) :]: v for k, v in ds.attrs.items() if k.startswith(key_prefix)
     }
     return ds
+
+
+def find_files_for_time_range(
+    product_type: ProductType,
+    time_range: Tuple[pd.Timestamp, pd.Timestamp],
+    get_files_for_path: GetFilesForPath,
+    dataset_filter: Optional[DatasetFilter] = None,
+) -> List[DatasetRecord]:
+    start, end = time_range
+
+    start_records = find_records_for_date(
+        product_type, start, get_files_for_path, dataset_filter
+    )
+    end_records = find_records_for_date(
+        product_type, end, get_files_for_path, dataset_filter
+    )
+
+    start_index = -1
+    for index, (_, _, start_end) in enumerate(start_records):
+        if start_end >= start:
+            start_index = index
+            break
+
+    end_index = -1
+    for index, (_, end_start, _) in enumerate(end_records):
+        if end_start >= end:
+            end_index = index
+            break
+
+    start_names = []
+    if start_index >= 0:
+        start_names.extend(start_records[start_index:])
+
+    # Add everything between start + start.day and end - end.day
+
+    start_p1d = (
+        pd.Timestamp(year=start.year, month=start.month, day=start.day, tz="UTC")
+        + _ONE_DAY
+    )
+    end_m1d = (
+        pd.Timestamp(year=end.year, month=end.month, day=end.day, tz="UTC") - _ONE_DAY
+    )
+
+    in_between_names = []
+    if end_m1d > start_p1d:
+        time = start_p1d
+        while time <= end_m1d:
+            in_between_names.extend(
+                find_records_for_date(
+                    product_type, time, get_files_for_path, dataset_filter
+                )
+            )
+            time += _ONE_DAY
+
+    end_names = []
+    if end_index >= 0:
+        end_names.extend(end_records[:end_index])
+
+    return start_names + in_between_names + end_names
+
+
+def find_records_for_date(
+    product_type: ProductType,
+    date: pd.Timestamp,
+    get_files_for_path: GetFilesForPath,
+    accept_record: Optional[DatasetFilter] = None,
+) -> List[DatasetRecord]:
+    path_pattern = product_type.path_pattern
+    name_pattern = product_type.name_pattern
+
+    year = date.year
+    month = date.month
+    day = date.day
+
+    prefix_path = path_pattern.format(
+        year=year,
+        month=f"0{month}" if month < 10 else month,
+        day=f"0{day}" if day < 10 else day,
+    )
+
+    records = []
+    for file_path in get_files_for_path(prefix_path):
+        parent_and_filename = file_path.rsplit("/", 1)
+        filename = (
+            parent_and_filename[1] if len(parent_and_filename) == 2 else file_path
+        )
+        m = re.match(name_pattern, filename)
+        if m is not None:
+            start = m.group("sd") + m.group("st")
+            end = m.group("ed") + m.group("et")
+            record = (
+                file_path,
+                pd.to_datetime(start, format=COMPACT_DATETIME_FORMAT, utc=True),
+                pd.to_datetime(end, format=COMPACT_DATETIME_FORMAT, utc=True),
+            )
+            if accept_record is None or accept_record(record):
+                records.append(record)
+
+    return sorted(records, key=lambda r: r[1])